Rename non-functional test to non-requested
I was very annoyed that the term non-functional is used. because non functional is per definition not possible.

this is so because function is defined as
- a process that a computer or a computer program uses to complete a task:
- the natural purpose (of something) or the duty (of a person)

The first definition says that the way a program completes a task… But who determines what the task is? 
because if a hacker needs to send one message to take down the system. so the function is handy for the hacker. 

The second definition says purpose. but here I could ask the same question. whose purpose?

And personally I read that someone has used the non-functional test and i get confused what it means. Furthermore, people use only the termInology to create a description. so some descriptions you receive that say the test has no function…. By using the term non-requested you could fix this.

you could also use the term the side effect test.